feat(Application) : 보너스번호  생성

+++ src/main/java/lotto/CreateBonusNumber.java
@@ -3,9 +3,9 @@ package lotto;
 import java.util.List;
 
 public class CreateBonusNumber {
-    public long getBonusNumber(List<Integer> winningNumber, String userInput){
+    public int getBonusNumber(List<Integer> winningNumber, String userInput){
         Validate checkBonusNumber = new Validate();
         checkBonusNumber.validateBonusNumber(winningNumber, userInput);
-        return Long.parseLong(userInput);
+        return Integer.parseInt(userInput);
     }
 }
